// skills-server/src/core/helpers/exception.enum.ts
export enum ExceptionEnum {
  CONTACT_N_F = 'Contact with this id not found',
  TODO_N_F = 'Task with this id not found',
  CONTACT_N_MARKED = 'This contact not marked for remove',
  USER_LOGIN_N_F = 'User with this login already exists',
  USER_L_N_F = 'User with this login not found',
  NO_ACCESS = 'Do not have access',
  USER_ROLE_N_F = 'User or Role not found',
  USER_N_F = 'User not found',
}